change to float point

+++ src/CubicSpline/CubicSpline2D.cpp
@@ -12,8 +12,8 @@ using namespace std;
 CubicSpline2D::CubicSpline2D() = default;
 
 // Construct the 2-dimensional cubic spline
-CubicSpline2D::CubicSpline2D(const vector<double> &x,
-                             const vector<double> &y) {
+CubicSpline2D::CubicSpline2D(const vector<float> &x,
+                             const vector<float> &y) {
     vector<vector<double>> filtered_points = remove_collinear_points(x, y);
     calc_s(filtered_points[0], filtered_points[1]);
     sx = CubicSpline1D(s, filtered_points[0]);
@@ -41,44 +41,44 @@ void CubicSpline2D::calc_s(const vector<double>& x,
 }
 
 // Calculate the x position along the spline at given t
-double CubicSpline2D::calc_x(double t) {
+float CubicSpline2D::calc_x(float t) {
     return sx.calc_der0(t);
 }
 
 // Calculate the y position along the spline at given t
-double CubicSpline2D::calc_y(double t) {
+float CubicSpline2D::calc_y(float t) {
     return sy.calc_der0(t);
 }
 
 // Calculate the curvature along the spline at given t
-double CubicSpline2D::calc_curvature(double t){
-    double dx = sx.calc_der1(t);
-    double ddx = sx.calc_der2(t);
-    double dy = sy.calc_der1(t);
-    double ddy = sy.calc_der2(t);
-    double k = (ddy * dx - ddx * dy) /
+float CubicSpline2D::calc_curvature(float t){
+    float dx = sx.calc_der1(t);
+    float ddx = sx.calc_der2(t);
+    float dy = sy.calc_der1(t);
+    float ddy = sy.calc_der2(t);
+    float k = (ddy * dx - ddx * dy) /
             pow(pow(dx, 2) + pow(dy, 2), 1.5);
     return k;
 }
 
 // Calculate the yaw along the spline at given t
-double CubicSpline2D::calc_yaw(double t) {
-    double dx = sx.calc_der1(t);
-    double dy = sy.calc_der1(t);
-    double yaw = atan2(dy, dx);
+float CubicSpline2D::calc_yaw(float t) {
+    float dx = sx.calc_der1(t);
+    float dy = sy.calc_der1(t);
+    float yaw = atan2(dy, dx);
     return yaw;
 }
 
 // Given x, y positions and an initial guess s0, find the closest s value
-double CubicSpline2D::find_s(double x, double y, double s0) {
-    double s_closest = s0;
-    double closest = INFINITY;
-    double si = s.front();
+float CubicSpline2D::find_s(float x, float y, float s0) {
+    float s_closest = s0;
+    float closest = INFINITY;
+    float si = s.front();
 
     do {
-        double px = calc_x(si);
-        double py = calc_y(si);
-        double dist = norm(x - px, y - py);
+        float px = calc_x(si);
+        float py = calc_y(si);
+        float dist = norm(x - px, y - py);
         if (dist < closest) {
             closest = dist;
             s_closest = si;
@@ -90,13 +90,13 @@ double CubicSpline2D::find_s(double x, double y, double s0) {
 
 // Remove any collinear points from given list of points by the triangle rule
 vector<vector<double>>
-CubicSpline2D::remove_collinear_points(vector<double> x, vector<double> y) {
+CubicSpline2D::remove_collinear_points(vector<float> x, vector<float> y) {
     vector<vector<double>> filtered_points;
     vector<double> x_, y_;
-    x_.push_back(x[0]);
-    x_.push_back(x[1]);
-    y_.push_back(y[0]);
-    y_.push_back(y[1]);
+    x_.push_back(static_cast<double>(x[0]));
+    x_.push_back(static_cast<double>(x[1]));
+    y_.push_back(static_cast<double>(y[0]));
+    y_.push_back(static_cast<double>(y[1]));
     for (size_t i = 2; i < x.size()-1; i++) {
         bool collinear = are_collinear(
             x[i - 2], y[i - 2],
@@ -106,8 +106,8 @@ CubicSpline2D::remove_collinear_points(vector<double> x, vector<double> y) {
         if (collinear) {
             continue;
         }
-        x_.push_back(x[i]);
-        y_.push_back(y[i]);
+        x_.push_back(static_cast<double>(x[i]));
+        y_.push_back(static_cast<double>(y[i]));
     }
     // make sure to add the last point in case all points are collinear
     x_.push_back(x.back());
@@ -118,9 +118,9 @@ CubicSpline2D::remove_collinear_points(vector<double> x, vector<double> y) {
 }
 
 // Determine if 3 points are collinear using the triangle area rule
-bool CubicSpline2D::are_collinear(double x1, double y1, double x2, double y2,
-    double x3, double y3) {
-    double a = x1 * (y2 - y3) +
+bool CubicSpline2D::are_collinear(float x1, float y1, float x2, float y2,
+    float x3, float y3) {
+    float a = x1 * (y2 - y3) +
                x2 * (y3 - y1) +
                x3 * (y1 - y2);
     return a <= 0.01;
