+++ src/Polynomials/QuinticPolynomial.cpp
@@ -5,8 +5,8 @@
 
 using namespace Eigen;
 
-QuinticPolynomial::QuinticPolynomial(double xs, double vxs, double axs,
-        double xe, double vxe, double axe, double t):
+QuinticPolynomial::QuinticPolynomial(float xs, float vxs, float axs,
+        float xe, float vxe, float axe, float t):
         a0(xs), a1(vxs) {
     a2 = axs / 2.0;
     Matrix3d A;
@@ -23,20 +23,20 @@ QuinticPolynomial::QuinticPolynomial(double xs, double vxs, double axs,
     a5 = x[2];
 }
 
-fixp_d QuinticPolynomial::calc_point(double t) {
+fixp_d QuinticPolynomial::calc_point(float t) {
     return a0 + a1 * t + a2 * pow(t, 2) + a3 * pow(t, 3) +
     a4 * pow(t, 4) + a5 * pow(t, 5);
 }
 
-fixp_d_d QuinticPolynomial::calc_first_derivative(double t) {
+fixp_d_d QuinticPolynomial::calc_first_derivative(float t) {
     return a1 + 2 * a2 * t + 3 * a3 * pow(t, 2) + 4 * a4 * pow(t, 3) +
     5 * a5 * pow(t, 4);
 }
 
-fixp_d_dd QuinticPolynomial::calc_second_derivative(double t) {
+fixp_d_dd QuinticPolynomial::calc_second_derivative(float t) {
     return 2 * a2 + 6 * a3 * t + 12 * a4 * pow(t, 2) + 20 * a5 * pow(t, 3);
 }
 
-fixp_d_ddd QuinticPolynomial::calc_third_derivative(double t) {
+fixp_d_ddd QuinticPolynomial::calc_third_derivative(float t) {
     return 6 * a3 + 24 * a4 * t + 60 * a5 * pow(t, 2);
 }
